fix: remove whitespaces before XML declaration
The envelope being sent out needs to start with the XML declaration '<?xml'.
According to the XML spec, no whitespace characters or new lines are allowed
before. Some routers seem to be very restrictive here, responding with error
502 if the XML declaration is indented with whitespaces.

The multi line string declaration for building the envelope needs to be
stripped therefore. Stripping the body_template as well for unification.

This fixes compatibility with FRITZ!Box WLAN 3170.